Restructuring - config and blocks were moved to config/ Things were changed according
