+++ srcs/CMakeLists.txt
@@ -2,14 +2,15 @@ project(webserv_lib)
 
 add_library(${PROJECT_NAME} STATIC
 	webserv.cpp
+	config/Blocks.cpp
+	config/Config.cpp
+	# config/ConfigSocket.cpp
 	http/HTTPHandler.cpp
 	http/HTTPParser.cpp
 	http/HTTPRequest.cpp
 	http/HTTPResponse.cpp
 	http/HTTPValidator.cpp
 	server/HTTPServer.cpp
-	server/Config.cpp
-	server/Blocks.cpp
 	utils/Logger.cpp
 	socket/Socket.cpp
 	socket/SocketAddress.cpp
